requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyYAML>=6.0
tqdm>=4.65.0
pillow>=10.0.0
opencv-python>=4.8.0
scipy>=1.10.0

# For baseline models
tsai>=0.3.4
pyts>=0.12.0

# For spectrograms
librosa>=0.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
